Test pcc gui and utils
